+++ src/renderer/components/wallet/assets/index.ts
@@ -1,5 +1,6 @@
 import { AssetDetails } from './AssetDetails'
 import { AssetsNav } from './AssetsNav'
 import { AssetsTableCollapsable } from './AssetsTableCollapsable'
+import { AssetsTotalBalances } from './AssetsTotalBalances'
 
-export { AssetDetails, AssetsNav, AssetsTableCollapsable }
+export { AssetDetails, AssetsNav, AssetsTableCollapsable, AssetsTotalBalances }